23.05.13 심플인포 승률 수정 및 오류 출력 확인

+++ src/main/java/com/example/demo/src/domain/user/service/UserService.java
@@ -125,7 +125,7 @@ public class UserService {
             int recentWinCount = historyDao.getRecentWinCount(userIdx);
             int recentLoseCount = historyDao.getRecentLoseCount(userIdx);
             int recentDrawCount = historyDao.getRecentDrawCount(userIdx);
-            int recentWinRate = (int) recentWinCount / 10;
+            int recentWinRate = recentWinCount / 10;
 
             return new UserSimpleInfo(
                     userInfo.getName(),
